# pages/api-docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "4.0",
    "title": "18 Candleriggs",
    "description": "This is the Candleriggs Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "name": "Superminds",
      "email": "contact@superminds.dev"
    }
  },
  "paths": {
...
    "/api/gallery/web-image-big/{collectionId}": {
      "post": {
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Upload a gallery web image big ",
        "description": "__How To__\n\n1. If you dont have a ObjectID hit an endpoint **[POST]/api/gallery** and get a valid ObjectID\n2. Fill the ObjectID.\n3. Select image to upload and click execute\n\n**Accepted image formats**\n1. png\n\n2. jpg\n\n3. jpeg\n\n **Accepted size and dimensions (width & height)** \n\nAccepted image size : under 500KB\n\nwidth: 630px\n\nheight: 600px\n\n**The below table defines the HTTP Status codes that this API may return**\n\n  | Status Code      | Description | Reason                                   |\n  | ---------------- | ------------| -----------------------------------------|\n  | 201              | CREATED          | Image created successfully.                       |\n  | 403              | FORBIDDEN | If user unauthenticated. |\n  | 404              | NOT FOUND   | If this collectionId not found.              |\n  | 422              | BAD REQUEST | If the collectionId is in an invalid format or image not correct. |\n  | 500              | SERVER ERROR| If server error encountered.             |\n",
        "operationId": "galleryWebImageBigPOST",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "description": "The Id of the collection to upload web image big",
            "schema": {
              "type": "string",
              "example": "63844c8bc1e969f2934f09f7"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "object",
                "properties": {
                  "webImageBig": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "msg": {
                      "type": "string",
                      "example": "Uploaded successfully"
                    },
                    "result": {
                      "allOf": [
                        {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "63a4522700f6553f757ee87d"
                            }
                          }
                        },
                        {
                          "$ref": "#/components/schemas/Gallery"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return400"
                }
              }
            },
            "description": "Data not received from user"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return403"
                }
              }
            },
            "description": "Forbidden: If user unauthenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "ObjectID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error."
          }
        }
      },
      "delete": {
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Delete the web image big of a gallery collection",
        "description": "__How To__\n\n1. If you dont have a collectionId hit an endpoint **GET /api/gallery** and get a valid collectionId. \n2. Fill the collectionId and execute.\n\n**The below table defines the HTTP Status codes that this API may return**\n\n  | Status Code      | Description | Reason                                   |\n  | ---------------- | ------------| -----------------------------------------|\n  | 200              | OK          | Image deleted successfully.                       |\n  | 403              | FORBIDDEN | If user unauthenticated. |\n  | 404              | NOT FOUND   | If this collectionId not found.              |\n  | 422              | BAD REQUEST | If the collectionId is in an invalid format. |\n  | 500              | SERVER ERROR| If server error encountered.             |\n",
        "operationId": "galleryWebImageBigDelete",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6399bf322f332f8babe66407"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "msg": {
                      "type": "string",
                      "example": "successfully deleted"
                    },
                    "result": {
                      "allOf": [
                        {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "637487643672cb71ea1444cd"
                            },
                            "smallWebImage": {
                              "type": "string",
                              "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
                            },
                            "bigWebImage": {
                              "type": "string",
                              "example": ""
                            },
                            "smallMobileImage": {
                              "type": "string",
                              "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
                            },
                            "bigMobileImage": {
                              "type": "string",
                              "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
                            }
                          }
                        },
                        {
                          "$ref": "#/components/schemas/GalleryWithoutImages"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": "success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return403"
                }
              }
            },
            "description": "Forbidden: If user unauthenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "BannerID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Invalid Id provided"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Return400": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "msg": {
            "type": "string",
            "example": "Bad Request: {message}"
          }
        }
      },
      "Return403": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 403
          },
          "msg": {
            "type": "string",
            "example": "Forbidden: {message}"
          }
        }
      },
      "Return404": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "msg": {
            "type": "string",
            "example": "Not Found: {message}"
          }
        }
      },
...
      "Return422": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 422
          },
          "msg": {
            "type": "string",
            "example": "Unprocessable Entity"
          },
          "errors": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ValidationError"
                }
              ]
            }
          }
        }
      },
      "Return500": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 500
          },
          "msg": {
            "type": "string",
            "example": "Internal Server Error: {message}"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Provide Password"
          },
          "param": {
            "type": "string",
            "example": "password"
          },
          "location": {
            "type": "string",
            "example": "body"
          }
        }
      },
...
      "Gallery": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GalleryWithoutImages"
          },
          {
            "type": "object",
            "properties": {
              "smallWebImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
              },
              "bigWebImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
              },
              "smallMobileImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
              },
              "bigMobileImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/11671713808569"
              }
            }
          }
        ]
      },
      "GalleryWithoutImages": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "Venue",
              "Events",
              "Food"
            ]
          },
          "name": {
            "type": "string",
            "example": "Turkish Eggs"
          },
          "uploadedDate": {
            "type": "string",
            "example": "2002-07-15",
            "format": "full-date",
            "externalDocs": {
              "description": "SeeMore",
              "url": "https://www.rfc-editor.org/rfc/rfc3339#section-5.6"
            }
          },
          "order": {
            "type": "number",
            "example": 0
          },
          "altText": {
            "type": "string",
            "example": "Girl in jacket"
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
...
    }
  }
}